Add images to podcast search
This also introduces an ugly hack for making sure that there are spaces between
the podcast boxes. The hack is that it fills up so that the number of "hits"
always is dividable by 4. This way, it is possible to use
`justify-content: space-between` in the css. It is hard to say what's worst,
hacky css or hacky js. I chose to do it in JS this time, might change my mind.